move lapack routines to lapack.cpp
